Update gui.py
This code change allow the use of comentes in Taipy string pages definitions, that using Markdown codes, like:  
page_login = """
/# The line below inset a button:
<|{Mat_txt[10][Idima_sel.id/* number of idiom = 1 to 8*/]}/* Mat_txt[10]=word "Login" in 8 idioms */ |button| active|on_action=confirma_login/* def confirma_login(state) is a python function called by button activation*/ |>\n """

before my clean procedure its becomes like:

page_login = """
<|{Mat_txt[10][Idima_sel.id]}|button| active|on_action =confirma_login|>\n """

For HTML codes the comments are made using /*    */ delimiters, for Markdown I define the same but including the use of a like python comment given by "/#" delimiter
So now two types of comments are allowed inside of the string that define the page using strings with Markdown code::
A- Line-ending comment: /# XXXX  =>  Need be used in the end of the line or in a complete line. All words before /# will be removed  
B- Comment inserted: /* XXXX  */    =>  Can be used inside the code. Only the text  /* XXXX  */   will be removed.
To do it modification, after user  call Gui(pages=pages).run() I do:
pages= remove_comment_from_pages(pages).
But if is only a page in pages is need to do
page1 = remove_comment(pages)
pages = {"/": page1}

The file _comment.py contain the rotines to remove the comment from page.
I teste it in whidows10 python 3.11 and Linux Debian 11, Python 3.11 and works ok. 
If this improvement on Taipy be accepted we need change the Taipy documentation to explain this new feature, and so many Taipy examples can be inprovide using comment.