Create _comment.py
This Python file contain two rotines that allow the use of comentes in Taipy peges that using Markdown codes. 
For HTML codes the comments are made using /*    */ delimiters, for Markdown we define the same but including the use of a like python comment given by /# delimiter.

The function remove_commen allows User to do comments in Markdown codes in order to facilitate understanding and documentation of what was developed.
Everything after /# until the end of the line will be removed.
Only the part that falls within the delimiters (/* */) will be removed.

remove_comment_from_pages funcition remove the commente of a pages_dict by call remove_commen page by page.